pyproject.toml
[project]
name = "firstproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mr. Fi$h", email = "77742137+MuhammadAsadkhalid@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "generativeai>=0.0.1",
    "google-generativeai>=0.8.5",
]

[project.scripts]
firstproject = "firstproject:main"

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"
